// frontend/src/api/generated/models/indicatorSeries.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * vibe-quant API
 * OpenAPI spec version: 0.1.0
 */
import type { IndicatorSeriesParams } from './indicatorSeriesParams';
import type { IndicatorSeriesPoint } from './indicatorSeriesPoint';

export interface IndicatorSeries {
  name: string;
  output_name: string;
  indicator_type: string;
  display_label: string;
  pane: string;
  params: IndicatorSeriesParams;
  data: IndicatorSeriesPoint[];
}